Added the invoice number issue fixed  and commit on 25-04-2025

+++ invoice_generator_tool/src/main/java/com/nithra/invoice_generator_tool/model/InvoiceAddedList.kt
@@ -73,7 +73,7 @@ class InvoiceAddedList {
 
             @SerializedName("paid_amt")
             @Expose
-            var paidAmt: Int? = null
+            var paidAmt: Double? = null
 
             @SerializedName("due_date")
             @Expose
