+++ invoice_generator_tool/src/main/java/com/nithra/invoice_generator_tool/model/InvoiceGetInvoiceList.kt
@@ -50,7 +50,7 @@ class InvoiceGetInvoiceList {
 
     @SerializedName("paid_amt")
     @Expose
-    var paidAmt: Int? = null
+    var paidAmt: Double? = null
 
     @SerializedName("due_date")
     @Expose
